Renames Pie components to PieChart
