--- packages/component-library/src/Pie/Pie.css
@@ -1,17 +1,0 @@
-/* Pie & Donut visualization */
-
-.donut-title {
-  font-family: 'filson-soft', sans-serif;
-  font-size: 1.5rem;
-  font-weight: 300;
-  color: #220f25;
-  fill: #220f25;
-}
-
-.donut-subtitle {
-  font-family: 'filson-soft', sans-serif;
-  font-weight: 300;
-  color: #706371;
-  fill: #706371;
-  text-transform: uppercase;
-}
