+++ packages/component-library/src/index.js
@@ -5,7 +5,7 @@ export { default as Button } from './Button/Button';
 export { default as StoryCard } from './StoryCard/StoryCard';
 export { default as Chart } from './Chart/Chart';
 export { default as ChartData } from './ChartData/ChartData';
-export { default as Pie } from './Pie/Pie';
+export { default as PieChart } from './PieChart/PieChart';
 export { default as Slider } from './Slider/Slider';
 export { default as Footer } from './Footer/Footer';
 export { default as ScrollToTop } from './ScrollToTop/ScrollToTop';
